reestruturado para melhor leitura
