+++ main.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include "main.h"
 #include "fila.h"
+#include "no.h"
 
 int main() {
 	return 0;
