+++ no.h
@@ -1,0 +1,7 @@
+typedef struct No no;
+struct No {
+	int item;
+	no *proximo;
+};
+
+no* criaNo(int item);
